update login button
